Add test for global write check (mutation inside function call)
Summary:
Test that checks if the global write check detects writes to global variables
passed as arguments to a function call.

The checker doesn't detect the write to the global variable `$bar` is mutated inside the `plus_one` function, `$fun` anonymous function and `Test::plus_two` method, as can be seen in the output. This test case, therefore,
demonstrates a false negative.

We will need to update the expected output when this case is handled correctly.

Differential Revision: D36519873

fbshipit-source-id: 242750aacee0f2ad91c0d6a73eb6cb8e22bf597b