HPHP::AutoloadMap::getTypeOrTypeAliasFile()
Summary:
Hack doesn't disambiguate references to types and type aliases. In a snippet like `function f(Foo $foo);`, we don't know if `Foo` is the name of a type (i.e. a class/interface/trait/enum), or if it's an alias to a type. The only way to know is to try loading `Foo`.

Today, we do this by first trying to load `Foo` as a type. If `Foo` is actually a type alias, the autoloader will just fail. We'll then try to load `Foo` as a type alias. This is wasteful. We *don't care* (yet) whether `Foo` is a type or type alias - we need to load its definition either way.

This diff adds a new mode, which for now is called `TypeOrTypeAlias`. It threads it through all our autoload maps.

**Note for OSS:** `autoload_set_paths()` accepts a callback, which returns a string representing the kind of symbol being loaded. If the symbol was a type, the kind string would be `"class"`. If the symbol was a type alias, the kind string would be `"type"`. Now, the kind string is `"type"` whether the symbol is a type alias or class. This is compatible with the WWW codebase, but may break compatibility in open source.

Reviewed By: edwinsmith

Differential Revision: D36015326

fbshipit-source-id: 1c8a7f1d1b25e7f888120db375a6ee83a324cc5a