Add APCHandle::toLocalMayRaise helper
Summary:
We can't make lazy APC dicts or lazy APC objects if there's a chance that toLocal() on an element or property will raise. Doing so would cause us to move an exception from apc_fetch to some field access, which is a user-visible semantics change.

To protect against these cases, we add a conservative implementation of toLocalMayRaise() to APCHandle, and implement it recursively when constructing any APC aggregates. For APCTypedValue cases - the most common "leaf" type - this method can trivially return true.

Reviewed By: ricklavoie

Differential Revision: D32707060

fbshipit-source-id: 91b71c3fc40a4d901c5b0fee41a49dc35626c780