ignore --using-mini-state option
Summary:
Currently, all our invocations of `hh_server` for indexing use this option "--config use_mini_state=true", but it serves no purpose and is confusing.

The option which determines running the indexer is `--write-symbol-info`.  The indexer doesn't make any use of `use_mini_state` but if the option isn't set, it prevents the indexer from running (by going to a different init mode).

The new behavior is to ignore `use_mini_state` if `--write-symbol-info` is set.

Reviewed By: pepeiborra

Differential Revision: D36511644

fbshipit-source-id: 56074343092b41ac07f82f1db065cb8e03a8d3e3