add support for class constants as bespoke sources
Summary:
A source for type structures can be a class constant. This diff adds class constants for type structures as a bespoke source: `LocationType::TypeConstant`. Class constants are being used from the `clsCnsGet` method or in irgen, this diff only covers the `clsCnsGet` case.

The method `profileArrLikeClsCns` is called after an array-like class constant has been resolved.

Modified some IR instructions in specification to have `LA` in order to access the bespoke layouts.

Reviewed By: oulgen

Differential Revision: D36268680

fbshipit-source-id: bba50a962eb51bdd86e5b8122568799804424235