Use Local instead of ParamName in HHBC and IR
Summary:
A while ago we converted hhbc_ast Locals from a named/unnamed enum to
a simple typesafe integer. At the time, we missed ParamName, which had the
same enum representation. And when we started macro-generating everything,
this required a few `replace_imm()` fixups.

Fix it by deleting ParamName and removing the fixups. One of them revealed
an an inconsistency in opcodes.h:  VerifyOutType takes a param id as IVA,
unlike other opcodes which take param ids as ILA. I changed it to ILA to match,
and added missing docs for VerifyOutType.

As a side effect this removes a few more arena string allocations.

Reviewed By: aorenste, kavoor

Differential Revision: D36627095

fbshipit-source-id: 18bc4d87f9cf84cabbf3eec36cf94813cb762f7d