Migrate `usize` integer to explicit `u64` for positional information (#9014)
Summary:
Once upon a time https://hacklang.org had an in-browser playground where people could experiment with Hack's typechecker. The in-browser playground was based on a transpiled-to-JS version of the OCaml source.

As the project migrates to Rust I'm interested in creating something similar with WASM, using the much-more-polished `wasm-pack` utility for Rust. One encumbrance is the platform requirements — Rust `usize` integers are treated as `u32` inside the WASM output, which breaks AST position information.

This PR attempts to fix that by being explicit about all numerical operations. On 64-bit platforms — all but WASM — these conversions should be no-ops, erased at compile time, since `usize` and `u64` are functionally identical.

This PR omits similar changes in the original ML source code on the understanding that the ML code will be phased out eventually, and that side of things currently works as expected.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9014

Reviewed By: Wilfred

Differential Revision: D34532680

fbshipit-source-id: cca5b34ea2436633042061ef417aaaefbbf42b5b