Initial Assembler Interface
Summary:
HackC CLI now has a shell interface of an "assemble" option. Specify which .hhas file to assemble into its HackCUnit representation. Because that representation is not human-readable assemble uses bytecode_printer to print the hhas representation of the HackCUnit (either to stdout or a specified output file).
Once assemble works the purpose will be to check that the input file and output are identical. In future iterations the underlying assemble function will be in its own crate like compile.
Currently assemble just returns a default HackCUnit, and that's what's printed out (along with metadata on the chosen file).

Reviewed By: aorenste

Differential Revision: D36630892

fbshipit-source-id: c134cba6e7e33644c35dfce1f256fc31b517c11d