log total = cgroup memory_current+memory_swap_current
Summary:
This diff requires explanation. There are two underlying measures:
* `/sys/fs/cgroups/<CGROUP/memory.current` - the total amount of physical RAM that's being used at the moment
* `/sys/fs/cgroups/<CGROUP>/memory.swap.current` - the total amount of swap that's being used at the moment

For instance, a cgroup might start out at current=2gb swap.current=0, and then transition to current=1gb swap.current=1gb, and it'd be using the exact same amount of data, but the OS decided to swap some of it out.

It's clear that the sum of the two, "current + swap.current", is the best measure of the overall burden imposed by hh_server. And indeed we've observed that (1) the best memory correlate of fanout / job-size is the sum "current + swap.current", (2) the best memory correlate of OOMing is when "current + swap.current" has become too big. That's not surprising: hh_server will grow its overall burden, and the OS will decide to swap some out if there's no room in physical ram, and then eventually when swap runs out too then we'll get an OOM.

It's also clear that measuring how "current" has changed since the start of the hh_server doesn't tell us how hh_server's burden has changed. (1) If something else is running concurrently, it tells us how the OS is weighting hh_server vs the other thing. (2) If we reach peak physical RAM but continue to impose burden, then "current" will remain steady even as the burden increases.

Here's another piece of background. We log cgroup information in several places.

**hh_server_cgroups scuba table**
* total_relative_to (memory.current when hh_server started)
* total_start (memory.current when the current "phase" started, minus total_relative_to)
* total (memory.current now at the end of the current "phase", minus total_relative_to)
* secs_at_total_gb (an array, showing for each gb, how long has memory/current minus total_relative_to been at this gb)
* secs_above_{20,25,...}gb_total (similar)
* totalswap_relative_to (as above but memory.swap.current)
* totalswap_start (as above but memory.swap.current)
* totalswap (as above but memory.swap.current)

**Hh_logger.log**
* total (memory.current now minus total_relative_to)
* hwm (for the current phase, what is the hwm of memory.current minus total_relative_to)

**TYPE_CHECK_END scuba event "data.typecheck.cgroup" json**
* relative_to_initial (same as total_relative_to)
* start_bytes (memory.current minus total_relative_to at the start of the typecheck)
* end_bytes (memory.current minus total_relative_to at the end of the typecheck)
* hwm_bytes (hwm of memory.current)

## WHAT THIS DIFF DOES

From the above description, it's pretty much useless for us to be logging memory.current for all these things. It's also misleading for us to be calling it "total".
1. This diff changes the meaning of "total" to now mean "memory.current + memory.swap.current"
2. This diff removes the totalswap entries in hh_server_cgroups scuba table. They don't make sense given the first change.

## DISCUSSION

It would have given us more data if we recorded raw information about memory.current and memory.swap.current, and left it to scuba derived queries to sum them up together. But (1) that would just have been a load of work to do, (2) I'm worried that people might have looked at memory.current and not done the maths, (3) it'd just multiply the number of lines of code too much.

Differential Revision: D36689768

fbshipit-source-id: b088392d13895cede3938211fa42768c1fe06bc3