Fix closure DCE bug
Summary: The point of `DecReleaseCheck` is to make the `DecRef` shallow so we should never be replacing it with a `DecRef` (or in this case a `DecReleaseCheck`) of the may-move sources of the killed instruction. The closure does still own the captured `$this`, which will need to be `DecRef`ed in the case of `ConstructClosure`.

Reviewed By: ottoni

Differential Revision: D36685097

fbshipit-source-id: 81eaf97ac5938240cadae4a0f7ed288e42d23de1