Add functions to 'force' solving of tyvars
Summary: Adds functions to 'force' a solution for a type variable. `force_solve` will first try to find a concrete type common to the upper and lower bounds and solve to it. If there is no such type, we bind to the (union of) the lower bound completely ignoring the observed variance of the tyvar

Differential Revision: D36416603

fbshipit-source-id: d3f1047db7556221923b94ee66b814f802ce2493