Phrase upper and lower bounds solely in terms of subtype constraints
Summary:
The OCaml implementation has the notion of an `InternalTy` which is the sum of `locl_ty` and `constraint`. The main purpose of this is to allow quick calculation of the new propositions which should be processesed under transitivity of `<:`. However, the type introduces quite a bit of noise and doesn't seem to be strictly necessary.

This diff removes our corresponding `CstrTy` and makes our upper and lower bounds `HashSet<Ty<R>>`

We will reintroduce (non-subtype) constraints in a subsequent diff

Differential Revision: D36623411

fbshipit-source-id: 394d1d3fe091e14b170a2d811e0bf4b8bcc8300e