Shrink HHBBC::Type to 24 bytes
Summary:
We can shrink DCls/DObj down to 8 bytes by encoding the ctx bit and
type tag into the pointer directly. DWaitHandle can be shrunk to 8
bytes by storing a copy_ptr to the DWaitHandle, not having a copy_ptr
in the DWaitHandle.

While we're at it, DCls and DObj are identical, so stop copying the
logic and use the same type for both.

Reviewed By: oulgen

Differential Revision: D36782922

fbshipit-source-id: 6ae0e0382e78182195bfd8e46b8fe831ef278b58