Fix unhelpful "X derives from X" lint text in Hack
Summary:
update

"Lint[5628] This as assertion will always succeed and hence is redundant. The expression on the left is an instance of MyFoo. It is always an instance of MyFoo because MyFoo derives from MyFoo."

to

"Lint[5628] This as assertion will always succeed and hence is redundant. The expression on the left is an instance of MyFoo."

Reviewed By: Wilfred

Differential Revision: D36784803

fbshipit-source-id: cbfa35c815fe2905984efa0f6cef187638b5136f