Better display of definition information on properties and methods
Summary:
The "Full name:" text is redundant with the method name in the signature, and it's shown right at the end of the hover info, whereas the signature is shown at the beginning.

{F735780732}

It also doesn't show on properties, and it's particularly confusing when methods aren't found.

{F735781493}

Instead, include information on where a method or property is defined as part of the signature.

Reviewed By: hgoldstein

Differential Revision: D36686137

fbshipit-source-id: 1e2c076063e438e79d7bce29d86103c818d56fc6