update prio to a newer version
Summary:
As per title. We update libprio to be the same version used by ISRG experimental implementation. The new version has serialization/deserialization supported.

Deleting the existing thrift service implementation since the Server::<Field32>::new API changes and we will be implementing the endpoints based on the privacy-preserving specification (https://github.com/ietf-wg-ppm/draft-ietf-ppm-dap).

We will sync up with ISRG to deal with the wasm compatibility issue after the backend is set up, and the inter-operability is tested out.

Reviewed By: dtolnay

Differential Revision: D36520732

fbshipit-source-id: 30474b6f2ac7df6b22b861f657316cb6072014e3