Remove push/pop_local_changes from FileProvider trait
Summary:
The push/pop_local_changes methods don't really belong in FileProvider. Only hh_server is interested in having the ability to push/pop local changes, and we really shouldn't be giving the (Rust) typechecker permission to invoke these methods anyway.

Instead, follow the same pattern used for the naming table in previous diffs, and for shallow and folded decls later in this stack: have HhServerProviderBackend hold on to a backing ChangesStore (in order to invoke push/pop_local_changes on it), and use that as the storage for the FileProvider.

Differential Revision: D36725433

fbshipit-source-id: d53529a4f96bac23012438bdc3b9170677bdaef9