Simplify FileProvider trait, prepare to connect hh_server's Rust FileProvider to sharedmem
Summary:
FileProvider should provide a minimal API for the typechecker. Since the typechecker only needs to read files, FileProvider should only provide a means of reading them (just as NamingProvider and DeclProvider only provide read methods, and no methods to remove invalidated symbols after disk changes--that behavior is handled by mutating the underlying store). The distinction between Disk and Ide files also doesn't need to be known by the typechecker, so we can lift FileType up into hackrs_provider_backend.

The current implementation of FileProvider is subtly wrong: in hh_server, when computing decls, if a file doesn't exist on disk, we silently behave as if the file did exist, but had no contents (i.e., File_provider.get_contents returns `Some("")`). Our Rust_provider_backend.File.get_contents function behaves this way, but when computing decls, FileProvider doesn't! It returns an error.

In order to preserve this error-returning behavior for use cases that want it (like our tests, and possible future tools), but exclude it in the FileProvider implementation used in hh_server, this diff moves the empty-string behavior to the hh_server-specific FileProvider implementation (i.e., the impl of FileProvider for the `FileProviderWithChanges` struct).

When it comes time to modify hh_server's Rust_provider_backend to use sharedmem, we only need substitute in a sharedmem store where this diff adds a `TODO: make this sharedmem` comment.

Differential Revision: D36725434

fbshipit-source-id: c1ce0412e577b35560572edbe01008b62cd4a268