Fix mocked class and const prop calculation involving interfaces in HHBBC
Summary:
When index calculates which classes could have sub-classes
with mocked classes or const props, it was just walking the parent
list. This isn't correct when dealing with interfaces because those
will be "parents" but won't be on the parent list. The correct way to
walk each class' subclass list and calculate it that way. This
probably doesn't cause any issues right now because we pessimize
interfaces so much.

Reviewed By: paulbiss

Differential Revision: D36819950

fbshipit-source-id: fc82790defba7acfa84f9025a97f85356305604b