Add a post-autoload listener callback that `AutomockTest` can use
Summary:
In order for Decls in Compilation to succeed, we need to kill userland autoloading and remove `require_once` from the language.

But a few thousand of the tests in WWW are `AutomockTest`s. `AutomockTest` instruments the autoloader:

https://www.internalfb.com/code/www/[dfd007d43d560fe88ad5c7a63fcf1439acc75ea1]/flib/intern/testinfra/automock/Automocker.php?lines=64-72

`Automocker::autoload` calls `Automocker::loadFile`:

https://www.internalfb.com/code/www/[dfd007d43d560fe88ad5c7a63fcf1439acc75ea1]/flib/intern/testinfra/automock/Automocker.php?lines=253-280

DavidSnider pointed out that, even with native autoloading, this code could still work. Instead of overriding the autoloader, we could implement a listener callback that runs after autoloading a unit. This would allow us to keep using AutomockTests in a native autoloading world.

**But** it would mean autoloading is observable, and it would mean autoloading would involve VM re-entry. From my talks with paulbiss, I got the impression that observable autoloading will block Decls in Comp (or some other optimization?). From my talks with jano, I got the impression that VM re-entry on autoloading would block Decls in Comp (or some other optimization?) But I'm not really sure what gets blocked or why. This diff is my way of asking.

Reviewed By: edwinsmith

Differential Revision: D35763819

fbshipit-source-id: d8a546710935d788956ef209bfe7acac3a8ced82