Add flag to "hh --refactor" to check if sound dynamic
Summary:
Make ```hh --refactor-check-sound-dynamic``` to check if the hack server is currently sound dynamic. This enables a user to check whether sound dynamic is being used, and it also gates information about sound dynamic refactoring behind a hidden flag.

This flag is not visible when calling ```hh --help``` or ```hh -help```, since sound dynamic is still an experimental feature.

Reviewed By: Wilfred

Differential Revision: D36756048

fbshipit-source-id: f6ad1631a5ec19277a3639bf1945f8e228b8b9b6