Translate Coeffects
Summary:
Translate Coeffects in HackC Translator.

I believe there may be a better way to share the static coeffects definition between HHVM and HackC using cxx. I tried first looking at the HHVM definition, but in HHVM, we represent Coeffects as a set of static strings. I'm not sure cxx supports sharing static strings. In HackC, I tried exposing a `to_string()` method for the `Ctx` enum. This was difficult because cxx requires custom types to be declared in the same module as the cxx_bridge, resulting in a more disorganized coeffects module in Rust.

I decided this approach of just adding a case-switch in c++ was the simplest approach for now.

Differential Revision: D36816360

fbshipit-source-id: 3e318724d880eca74331162057f336a0adb28c38