add type alias as bespoke source
Summary:
This diff adds type alias as a bespoke source: `LocationType::TypeAlias`. When we want to access the resolvedTypeStructure field in a type alias, we check if the array already has a profile, and if so, we set the array layout to be the bespoke layout before returning. Otherwise, we try to profile the array and return the type structure wrapping a logging array.

Added some small helper functions in type-alias

Note: this diff only handles cases where the type alias is pre-resolved. `resolvedTypeStructure` is only set if the type alias is pre-resolved.

Reviewed By: kavoor

Differential Revision: D36427222

fbshipit-source-id: bb1614b73890afaed985cd73feffcf71eac06a79