StringUtil::Explode: check limit before String::find
Summary:
By inspecting this code, I found that the result of `pos = input.find(delimiter, pos)` isn't used when `--limit > 1` is false.
This code has no side effects (and `pos` is not used after the `while` loop) and should never fail, so it is safe to reverse the boolean expression.  When the limit is actually reached, this will save some computation cost because `input.find(...)` is skipped.

Reviewed By: mofarrell

Differential Revision: D36290656

fbshipit-source-id: 2757a067e0dd49ab6e02fa0377aecb4c52f0ee3e