Remove redundant sha1 hashing
Summary:
hackc_facts_to_json_cpp_ffi() serializes a FactsResult to JSON.
FactsResult already contains a sha1sum, so we don't need to pass
in the source text and hash it one more time.

hackc_decls_to_facts_cpp_ffi() used to take a source_text argument
and compute the sha1 to include in FactsResult. But in the normal
flow of this API, the caller already computed a sha1, so use it instead
of recomputing it again.

Reviewed By: jthemphill

Differential Revision: D36875636

fbshipit-source-id: 1b7c4ce844975fc8f329638c668a4b888a95e81e