Remove ProviderBackend struct
Summary: This struct was added with the intent of making it a common interface for accessing providers, but different use cases will need different subsets of the set of all Provider traits. Typechecking, for instance, needs FileProvider (to parse files into full ASTs) and FoldedDeclProvider, but shouldn't be allowed access to a ShallowDeclProvider (because a fully-populated FoldedDeclProvider contains all the same information, so in use cases that compute all folded decls ahead of time like Zoncolan, maintaining a shallow classes store is a waste of memory). Tools that handle decls need decl providers, but no direct access to a file provider or naming provider.

Differential Revision: D36826459

fbshipit-source-id: e24f73bf2942e051040f979eef145351b356bbb9