Translate Coeffects (#9103)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/9103

Translate Coeffects in HackC Translator.

I believe there may be a better way to share the static coeffects definition between HHVM and HackC using cxx. I tried first looking at the HHVM definition, but in HHVM, we represent Coeffects as a set of static strings. I'm not sure cxx supports sharing static strings. In HackC, I tried exposing a `to_string()` method for the `Ctx` enum. This was difficult because cxx requires custom types to be declared in the same module as the cxx_bridge, resulting in a more disorganized coeffects module in Rust.

I decided this approach of just adding a case-switch in c++ was the simplest approach for now.

Reviewed By: edwinsmith

Differential Revision: D36816360

fbshipit-source-id: 814b7e00faa230d3c93a8d32af60d6887a603319