Targeted pessimisation for return types
Summary:
Extend logging to produce a line marked with `ret` and an empty name (cf `param` with parameter name) when return type is not enforced.

Update forward pessimiser to make use of the logging and produce `~` on return types only if `ret` is present for that line.

Also update post-visit logging by the pessimiser itself to report uses of ~ etc in return types.

Reviewed By: vassilmladenov

Differential Revision: D36838683

fbshipit-source-id: 30bdaa04f81107267fe675a40768b2692f7f5376