Use RAII guard for LockFreePtrWrapper locks and unlocks
Summary:
The LockFreePtrWrapper API is tricky to get right in the
presence of exceptions. Once you lock it, you need to manually unlock
it on all exit paths, including exceptions. Rather than have every
user roll their own solution (usually with SCOPE_EXIT), have
lock_for_update() return a ScopedLock object. The lifetime of the
ScopedLock encapsulates the lock. When it is destroyed the lock is
automatically released. For update_and_unlock, you can call a method
on the ScopedLock to update and unlock (which prevents the ScopedLock
from again unlocking in the dtor).

Convert all users over to the new API.

Reviewed By: mofarrell

Differential Revision: D36904518

fbshipit-source-id: c4760eb91a5ab0a2dd2c879823e41d55bf2da175