Use original upper bound for native function signatures (and typecheck ext_apc)
Summary:
When we compile systemlib functions, we do a small post processing step on native function signatures.

Each function has a return type constraint that HHVM uses to do runtime verification and for native functions, determine what the C++ signature *ought* to look like. Previously the logic was something like:
* If we're unable to create a user type or there was no return type hint to begin with, assume that this function is `void` (to align to C++ void)
* If the return type hint was `mixed` or `callable`, assume this function has no constraint (to align to `HPHP::Variant`)
* Otherwise, just re-use the user type as the type constraint, overwriting the original constraint determined by HackC

That last bullet is ... annoying: it allows us to properly declare native functions that return `void` as being implemented with C++ void, but it also means naively using `shape` in a return type doesn't work: it's interpreted as a user type `HH\shape`, which aligns with `HPHP::Variant`.

I've changed this last case to special case `void` and otherwise reuse the previously determined type constraint. This means `shape` will map to `darray`, which maps to `HPHP::Array`, which is what we want.

This was blocking typechecking `ext_apc` nicely, not to mention the functions already "typechecked" that could make use of this.

Reviewed By: aorenste

Differential Revision: D36842824

fbshipit-source-id: c8dc47c2966115946939b6883c5a23e2c42cf21f