(One more time) Replace Zero-sized type in ffi
Summary:
Was running into a bug on `memoize-async-eager.php`, and couldn't figure out why the value for local range was incorrect.

It turns out, the issue was related to using the zero-sized type, DUMMY in our C-header. In Rust, zero-sized types are 0 bytes.
In C, empty structs are undefined, but there is a GNU extension which supports them and they have size 0.
In C++, they are 1-byte.

Therefore, it appears the HackC-Translator was accessing the wrong memory after the DUMMY field.

Let's not make dummy zero-sized so C and C++ agree on header size.

Reviewed By: edwinsmith

Differential Revision: D36987812

fbshipit-source-id: 17b7893e707a01100f5a0a2ffc5692db4a1f1a19