Allow (dynamic & t) | u to be passed to SDT function expecting u
Summary:
For an SDT function whose signature is `<<__SupportDynamicType>> function foo(t $x):u`, we implement a special rule for function application that allows an argument of type `~t` to be passed, producing a result of type `~u`. This is justified semantically by considering the function to have an intersection type `(function(t):u) & (function(dynamic):dynamic)`.

Unfortunately the rule only fires for arguments whose type is exactly equivalent to `dynamic | t`. In particular, refinement can produce an intersection, for example, we might want to pass `(dynamic & nonnull) | t`. As above, this is legal, by considering `dynamic & nonnull <: dynamic` for the "dynamic" side of the intersection.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D37005487

fbshipit-source-id: e8727ea39faf96c684499e5a91131de9cb297470