Targeted pessimisation of properties
Summary:
Add support in hh for logging of pessimised properties. Also do this for parameters that are "promoted" to properties.

Support consumption of the log in `hh_pessimsation`.

Note that it might be nice in future to have separate logging controls (e.g. `pessimise.param` and `pessimise.prop`), but we need to be able to distinguish between a log that just happens to contain zero occurrences of a pessimised parameter, and one for which we want always-pessimise (i.e. non-targeting of parameters). Otherwise we need to keep logs and controls to `hh_pessimisation` in sync.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D36970484

fbshipit-source-id: 812f5e6bc87f1ae29432f67482cd0db090cab4c2