Add support for lazy APC objects
Summary:
In reviewing the results from the lazy APC dict optimization, kavoor, colavitam, and I found that laziness at the array level isn't helpful on many real workloads unless we also support lazily deserialized objects.

This diff is an initial version of lazy APC objects. Such an object can have APC handles for properties that are only localized on first read.

The key problem with this idea is that it could require us to pessimize all property access. However, we use a PGO-based trick to avoid this global pessimization. In the APC logic, we attempt to set a flag on a class before making a lazy object of that class, and during retranslateAll, we block sets of these flags. As a result, when generating optimized code for any object that has the flag unset, we can access its properties in the JIT with no lazy deserialization checks!

The upshot of this approach is that we force checks on the properties of any object that has been stored to APC during profiling, but to no other objects.

Reviewed By: ricklavoie

Differential Revision: D32440721

fbshipit-source-id: 3c64f86e9e58769eb97bd0bf6c97887f086c50e2