Detect addition/removal of __ConsistentConstruct attribute on class with implicit constructor
Summary: We incorrectly considered a class with no explicit constructor to be unchanged when the __ConsistentConstruct attribute was added or removed. It is changed, though, and uses of the constructor need to be rechecked, so look for changes to the class' constructor consistency even when sc_constructor is None (in the old and/or new version).

Reviewed By: jamesjwu

Differential Revision: D37049362

fbshipit-source-id: 19929ef9e8069bffa4f03763bd00c5a5c48e7160