Add FuncBuilder::rewrite_block
Summary:
Switches FuncBuilder to work on a borrowed version of Func instead of owned.

Adds FuncBuilder::rewrite_block() - rewriting a block in-place can be tricky.  Because of Rust ownership rules you need to temporarily swap out the block to be rewritten so the rewrite function can still access the Instruct and Func.  rewrite_block() localizes this process so passes can share that logic.

Reviewed By: edwinsmith

Differential Revision: D37006005

fbshipit-source-id: 61b2cb4e47a3a3d7f307c863d2f794a02d78e590