set the decl store to use local memory
Summary:
High-level goal: migrate clientLsp to folded decls. See attached task.

Migrating clientLsp to folded decls required folded decls to work with local memory since this is what clientLsp uses.

For now it's impossible to use folded decl with the local memory backend, because folded decls are hard-wired to use the Decl_store which is hard-wired to use shared memory.

This diff adds a folded decl cache to the local memory back-end, which corresponds to the shared memory heap for folded classes defined as Decl_heap.Class. This folded decl cache is then used to create and set set a decl store when we set a local memory backend (making Provider_backend depends on Decl_store).

Differential Revision: D36778492

fbshipit-source-id: e565900c5c37a4f0916ddbb6da22eeabd4134e66