Retune some inlining thresholds, take 2
Summary:
This is a retry of D36547559 (https://github.com/facebook/hhvm/commit/af760f6189d8812757b34e1cb20d75639c989a05), but dropping the change to HHIRInliningMaxDepth.

--

I noticed that the JIT was inlining a lot of cold code, and missing
opportunities in hotter code.  This diff retunes some thresholds to reduce the
inlining of colder code, and use the code-size budget to do more aggressive
inling of hotter code.

Reviewed By: oulgen

Differential Revision: D37063004

fbshipit-source-id: 0241c8f510cd6037eebae500040dbe79a573ff3d